1. Increase the threshold more for taylor bar case; 2. Add regression test for stretching case; 3. New source and test cases are moved to the core library.
