+++ src/shared/materials/inelastic_solid.h
@@ -22,11 +22,9 @@
  * ------------------------------------------------------------------------- */
 /**
  * @file 	inelastic_solid.h
- * @brief 	These are classes for define properties of elastic solid materials.
- *			These classes are based on isotropic linear elastic solid.
- * 			Several more complex materials, including neo-hookean, FENE noe-hookean
- *			and anisotropic muscle, are derived from the basic elastic solid class.
- * @author	Xiangyu Hu and Chi Zhang
+ * @brief 	These are classes for define properties of plastic solid materials.
+ * 			Several plastic materials including linear hardening, non-linear hardening and viscous plastic, are presented.
+ * @author	Xiaojing Tang, Dong Wu, and Xiangyu Hu
  */
 #pragma once
 
@@ -90,5 +88,77 @@ class HardeningPlasticSolid : public PlasticSolid
     virtual Matd ElasticLeftCauchy(const Matd& deformation, size_t index_i, Real dt = 0.0) override;
 
     virtual HardeningPlasticSolid *ThisObjectPtr() override { return this; };
+};
+
+/**
+ * @class NonlinearHardeningPlasticSolid
+ * @brief Class for plastic solid with nonlinear hardening
+ */
+class NonLinearHardeningPlasticSolid : public HardeningPlasticSolid
+{
+protected:
+	Real saturation_flow_stress_, saturation_exponent_;
+public:
+	/** Constructor */
+	explicit NonLinearHardeningPlasticSolid(Real rho0, Real youngs_modulus, Real poisson_ratio, Real yield_stress,
+		Real hardening_modulus, Real saturation_flow_stress, Real saturation_exponent)
+		: HardeningPlasticSolid(rho0, youngs_modulus, poisson_ratio, yield_stress, hardening_modulus),
+		saturation_flow_stress_(saturation_flow_stress), saturation_exponent_(saturation_exponent)
+	{
+		material_type_name_ = "NonLinearHardeningPlasticSolid";
+
+	};
+	virtual ~NonLinearHardeningPlasticSolid() {};
+
+	Real NonlinearHardening(Real hardening_parameter_pre)
+	{
+		return  (hardening_modulus_ * hardening_parameter_pre + yield_stress_
+			+ (saturation_flow_stress_ - yield_stress_) * (1 - exp(-saturation_exponent_ * hardening_parameter_pre)));
+	};
+
+	Real NonlinearHardeningDerivative(Real hardening_parameter_pre)
+	{
+		return  (hardening_modulus_ + saturation_exponent_ * (saturation_flow_stress_ - yield_stress_)
+			* exp(-saturation_exponent_ * hardening_parameter_pre));
+	};
+
+	/** compute the stress through deformation, and plastic relaxation. */
+	virtual Matd PlasticConstitutiveRelation(const Matd& deformation, size_t index_i, Real dt = 0.0) override;
+	/** compute the elastic part of normalized left Cauchy-Green deformation gradient tensor. */
+	virtual Matd ElasticLeftCauchy(const Matd& deformation, size_t index_i, Real dt = 0.0) override;
+
+
+	virtual NonLinearHardeningPlasticSolid* ThisObjectPtr() override { return this; };
+};
+
+/**
+ * @class ViscousPlasticSolid
+ * @brief Class for plastic solid with viscosity
+ */
+class ViscousPlasticSolid : public PlasticSolid
+{
+protected:
+	Real viscous_modulus_;
+	Real Herschel_Bulkley_power_;
+	const Real sqrt_2_over_3_ = sqrt(2.0 / 3.0);
+	StdLargeVec<Matd> inverse_plastic_strain_; /**< inverse of plastic right cauchy green strain tensor */
+
+public:
+	/** Constructor */
+	explicit ViscousPlasticSolid(Real rho0, Real youngs_modulus, Real poisson_ratio, Real yield_stress, Real viscous_modulus, Real herschel_bulkley_power)
+		: PlasticSolid(rho0, youngs_modulus, poisson_ratio, yield_stress), viscous_modulus_(viscous_modulus), Herschel_Bulkley_power_(herschel_bulkley_power)
+	{
+		material_type_name_ = "ViscousPlasticSolid";
+	};
+	virtual ~ViscousPlasticSolid() {};
+
+	virtual void initializeLocalParameters(BaseParticles* base_particles) override;
+	Real ViscousModulus() { return viscous_modulus_; };
+	/** compute the stress through deformation, and plastic relaxation. */
+	virtual Matd PlasticConstitutiveRelation(const Matd& deformation, size_t index_i, Real dt = 0.0) override;
+	/** compute the elastic part of normalized left Cauchy-Green deformation gradient tensor. */
+	virtual Matd ElasticLeftCauchy(const Matd& deformation, size_t index_i, Real dt = 0.0) override;
+
+	virtual ViscousPlasticSolid* ThisObjectPtr() override { return this; };
 };
 } // namespace SPH
