+++ tests/2d_examples/test_2d_stretching/regression_test_tool/BeamObserver_Position_Run_0_result.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<result>
+    <Snapshot_Element>
+        <Snapshot number_of_snapshot_for_local_result_="198" />
+    </Snapshot_Element>
+    <Result_Element>
+        <Particle_0 snapshot_0="~[0.0266728,0.00601972]" snapshot_1="~[0.0266738,0.00601636]" snapshot_2="~[0.0266746,0.00601092]" snapshot_3="~[0.0266785,0.00599679]" snapshot_4="~[0.0266804,0.00598398]" snapshot_5="~[0.0266817,0.00597223]" snapshot_6="~[0.026682,0.0059609]" snapshot_7="~[0.0266824,0.00595034]" snapshot_8="~[0.0266825,0.00593986]" snapshot_9="~[0.026683,0.00592988]" snapshot_10="~[0.0266833,0.00592022]" snapshot_11="~[0.0266835,0.00591072]" snapshot_12="~[0.0266837,0.00590126]" snapshot_13="~[0.026684,0.00589196]" snapshot_14="~[0.0266839,0.00588279]" snapshot_15="~[0.0266839,0.00587379]" snapshot_16="~[0.0266841,0.00586482]" snapshot_17="~[0.0266842,0.00585603]" snapshot_18="~[0.0266842,0.00584728]" snapshot_19="~[0.026684,0.00583841]" snapshot_20="~[0.0266839,0.00582974]" snapshot_21="~[0.026684,0.00582104]" snapshot_22="~[0.0266836,0.0058125]" snapshot_23="~[0.0266835,0.00580369]" snapshot_24="~[0.0266831,0.00579504]" snapshot_25="~[0.0266829,0.00578625]" snapshot_26="~[0.0266828,0.0057774]" snapshot_27="~[0.0266825,0.00576884]" snapshot_28="~[0.0266824,0.00576022]" snapshot_29="~[0.0266818,0.00575163]" snapshot_30="~[0.0266818,0.00574316]" snapshot_31="~[0.0266812,0.00573456]" snapshot_32="~[0.0266812,0.005726]" snapshot_33="~[0.026681,0.0057175]" snapshot_34="~[0.0266807,0.00570884]" snapshot_35="~[0.0266806,0.00570041]" snapshot_36="~[0.0266805,0.00569179]" snapshot_37="~[0.02668,0.00568332]" snapshot_38="~[0.0266794,0.0056749]" snapshot_39="~[0.0266792,0.0056663]" snapshot_40="~[0.0266788,0.00565777]" snapshot_41="~[0.0266783,0.00564944]" snapshot_42="~[0.0266781,0.00564092]" snapshot_43="~[0.0266779,0.00563243]" snapshot_44="~[0.0266777,0.00562403]" snapshot_45="~[0.0266775,0.00561558]" snapshot_46="~[0.0266773,0.00560718]" snapshot_47="~[0.0266771,0.00559878]" snapshot_48="~[0.0266767,0.00559037]" snapshot_49="~[0.0266763,0.00558196]" snapshot_50="~[0.0266759,0.00557347]" snapshot_51="~[0.0266754,0.00556503]" snapshot_52="~[0.0266752,0.00555662]" snapshot_53="~[0.0266749,0.0055482]" snapshot_54="~[0.0266744,0.00553977]" snapshot_55="~[0.0266742,0.0055313]" snapshot_56="~[0.0266742,0.0055227]" snapshot_57="~[0.0266739,0.00551411]" snapshot_58="~[0.0266738,0.00550554]" snapshot_59="~[0.0266737,0.00549699]" snapshot_60="~[0.0266736,0.00548826]" snapshot_61="~[0.0266734,0.00547955]" snapshot_62="~[0.0266731,0.0054709]" snapshot_63="~[0.0266728,0.00546226]" snapshot_64="~[0.0266726,0.00545355]" snapshot_65="~[0.0266725,0.00544481]" snapshot_66="~[0.0266726,0.00543601]" snapshot_67="~[0.0266724,0.00542734]" snapshot_68="~[0.0266722,0.00541855]" snapshot_69="~[0.0266721,0.00540968]" snapshot_70="~[0.026672,0.00540075]" snapshot_71="~[0.0266718,0.00539172]" snapshot_72="~[0.0266719,0.00538272]" snapshot_73="~[0.0266719,0.00537363]" snapshot_74="~[0.0266717,0.00536456]" snapshot_75="~[0.0266718,0.00535558]" snapshot_76="~[0.0266718,0.00534629]" snapshot_77="~[0.0266718,0.00533706]" snapshot_78="~[0.026672,0.00532778]" snapshot_79="~[0.0266719,0.00531853]" snapshot_80="~[0.026672,0.0053091]" snapshot_81="~[0.0266724,0.00529975]" snapshot_82="~[0.0266723,0.00529002]" snapshot_83="~[0.0266727,0.00528058]" snapshot_84="~[0.0266728,0.00527064]" snapshot_85="~[0.026673,0.00526087]" snapshot_86="~[0.0266733,0.00525099]" snapshot_87="~[0.0266737,0.00524109]" snapshot_88="~[0.0266741,0.00523112]" snapshot_89="~[0.0266745,0.005221]" snapshot_90="~[0.0266751,0.00521082]" snapshot_91="~[0.0266755,0.00520045]" snapshot_92="~[0.026676,0.00519005]" snapshot_93="~[0.0266766,0.00517941]" snapshot_94="~[0.026677,0.00516866]" snapshot_95="~[0.0266774,0.00515788]" snapshot_96="~[0.0266782,0.005147]" snapshot_97="~[0.0266791,0.005136]" snapshot_98="~[0.02668,0.00512488]" snapshot_99="~[0.0266809,0.00511366]" snapshot_100="~[0.0266817,0.00510233]" snapshot_101="~[0.0266825,0.00509063]" snapshot_102="~[0.0266834,0.00507887]" snapshot_103="~[0.0266843,0.00506678]" snapshot_104="~[0.0266853,0.00505459]" snapshot_105="~[0.0266864,0.00504268]" snapshot_106="~[0.0266876,0.00503016]" snapshot_107="~[0.0266888,0.00501779]" snapshot_108="~[0.0266901,0.00500521]" snapshot_109="~[0.0266915,0.0049925]" snapshot_110="~[0.0266927,0.00497983]" snapshot_111="~[0.0266941,0.00496688]" snapshot_112="~[0.0266954,0.00495361]" snapshot_113="~[0.0266966,0.00493989]" snapshot_114="~[0.0266979,0.00492632]" snapshot_115="~[0.0266995,0.00491284]" snapshot_116="~[0.0267011,0.00489919]" snapshot_117="~[0.0267029,0.00488522]" snapshot_118="~[0.0267046,0.00487132]" snapshot_119="~[0.0267063,0.00485745]" snapshot_120="~[0.0267078,0.0048428]" snapshot_121="~[0.0267093,0.00482836]" snapshot_122="~[0.026711,0.00481351]" snapshot_123="~[0.0267125,0.00479857]" snapshot_124="~[0.0267145,0.00478398]" snapshot_125="~[0.0267164,0.00476896]" snapshot_126="~[0.0267183,0.00475384]" snapshot_127="~[0.0267201,0.00473852]" snapshot_128="~[0.0267221,0.00472321]" snapshot_129="~[0.0267241,0.00470778]" snapshot_130="~[0.0267259,0.00469198]" snapshot_131="~[0.0267281,0.00467632]" snapshot_132="~[0.0267301,0.00466004]" snapshot_133="~[0.0267321,0.0046441]" snapshot_134="~[0.0267341,0.00462781]" snapshot_135="~[0.0267362,0.00461183]" snapshot_136="~[0.0267382,0.00459548]" snapshot_137="~[0.02674,0.00457867]" snapshot_138="~[0.0267422,0.00456174]" snapshot_139="~[0.0267444,0.00454486]" snapshot_140="~[0.0267464,0.00452813]" snapshot_141="~[0.0267486,0.00451076]" snapshot_142="~[0.0267509,0.00449406]" snapshot_143="~[0.0267532,0.00447702]" snapshot_144="~[0.0267556,0.00446002]" snapshot_145="~[0.0267581,0.00444276]" snapshot_146="~[0.0267604,0.0044257]" snapshot_147="~[0.0267631,0.00440847]" snapshot_148="~[0.0267655,0.00439079]" snapshot_149="~[0.0267675,0.00437325]" snapshot_150="~[0.0267698,0.00435506]" snapshot_151="~[0.0267723,0.0043375]" snapshot_152="~[0.0267748,0.00432002]" snapshot_153="~[0.0267773,0.00430196]" snapshot_154="~[0.0267799,0.00428398]" snapshot_155="~[0.0267823,0.00426554]" snapshot_156="~[0.0267848,0.00424762]" snapshot_157="~[0.0267877,0.00422979]" snapshot_158="~[0.02679,0.00421163]" snapshot_159="~[0.0267924,0.00419286]" snapshot_160="~[0.0267951,0.00417486]" snapshot_161="~[0.0267979,0.00415627]" snapshot_162="~[0.0268006,0.00413813]" snapshot_163="~[0.0268037,0.00411999]" snapshot_164="~[0.0268068,0.00410165]" snapshot_165="~[0.0268095,0.00408307]" snapshot_166="~[0.0268123,0.00406431]" snapshot_167="~[0.026815,0.00404538]" snapshot_168="~[0.0268182,0.00402666]" snapshot_169="~[0.0268211,0.00400769]" snapshot_170="~[0.0268241,0.00398892]" snapshot_171="~[0.0268271,0.00396965]" snapshot_172="~[0.0268294,0.00395061]" snapshot_173="~[0.0268322,0.00393144]" snapshot_174="~[0.0268356,0.00391315]" snapshot_175="~[0.026839,0.00389433]" snapshot_176="~[0.0268424,0.00387522]" snapshot_177="~[0.0268457,0.00385659]" snapshot_178="~[0.0268489,0.00383727]" snapshot_179="~[0.0268521,0.00381819]" snapshot_180="~[0.0268553,0.00379904]" snapshot_181="~[0.0268586,0.0037804]" snapshot_182="~[0.026862,0.00376105]" snapshot_183="~[0.0268654,0.00374162]" snapshot_184="~[0.0268686,0.0037224]" snapshot_185="~[0.0268723,0.00370325]" snapshot_186="~[0.0268761,0.0036845]" snapshot_187="~[0.02688,0.00366543]" snapshot_188="~[0.0268837,0.00364633]" snapshot_189="~[0.0268871,0.00362675]" snapshot_190="~[0.0268911,0.0036072]" snapshot_191="~[0.0268945,0.00358832]" snapshot_192="~[0.0268981,0.00356913]" snapshot_193="~[0.0269019,0.00354993]" snapshot_194="~[0.0269056,0.00353064]" snapshot_195="~[0.0269098,0.00351214]" snapshot_196="~[0.0269137,0.00349301]" snapshot_197="~[0.0269178,0.00347416]" />
+    </Result_Element>
+</result>
