+++ tests/2d_examples/test_2d_stretching/regression_test_tool/BeamObserver_Position_Run_2_result.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<result>
+    <Snapshot_Element>
+        <Snapshot number_of_snapshot_for_local_result_="198" />
+    </Snapshot_Element>
+    <Result_Element>
+        <Particle_0 snapshot_0="~[0.0266669,0.00601884]" snapshot_1="~[0.0266679,0.0060155]" snapshot_2="~[0.0266687,0.00601004]" snapshot_3="~[0.0266724,0.00599545]" snapshot_4="~[0.0266743,0.00598219]" snapshot_5="~[0.0266751,0.00597]" snapshot_6="~[0.0266755,0.00595837]" snapshot_7="~[0.0266758,0.00594762]" snapshot_8="~[0.0266759,0.00593693]" snapshot_9="~[0.0266763,0.00592683]" snapshot_10="~[0.0266768,0.00591697]" snapshot_11="~[0.026677,0.00590729]" snapshot_12="~[0.0266771,0.00589787]" snapshot_13="~[0.0266773,0.00588846]" snapshot_14="~[0.0266772,0.00587912]" snapshot_15="~[0.0266772,0.00587014]" snapshot_16="~[0.0266774,0.00586112]" snapshot_17="~[0.0266775,0.00585221]" snapshot_18="~[0.0266775,0.00584335]" snapshot_19="~[0.0266775,0.00583429]" snapshot_20="~[0.0266775,0.0058254]" snapshot_21="~[0.0266775,0.00581655]" snapshot_22="~[0.0266772,0.00580759]" snapshot_23="~[0.026677,0.0057989]" snapshot_24="~[0.0266764,0.00579018]" snapshot_25="~[0.0266763,0.00578136]" snapshot_26="~[0.026676,0.00577252]" snapshot_27="~[0.0266758,0.00576377]" snapshot_28="~[0.0266756,0.00575502]" snapshot_29="~[0.0266751,0.00574623]" snapshot_30="~[0.0266749,0.00573762]" snapshot_31="~[0.0266744,0.00572882]" snapshot_32="~[0.0266742,0.00572022]" snapshot_33="~[0.0266739,0.00571159]" snapshot_34="~[0.0266735,0.00570295]" snapshot_35="~[0.0266733,0.00569446]" snapshot_36="~[0.0266731,0.00568586]" snapshot_37="~[0.0266725,0.0056772]" snapshot_38="~[0.0266719,0.00566858]" snapshot_39="~[0.0266716,0.00566004]" snapshot_40="~[0.0266711,0.00565132]" snapshot_41="~[0.0266706,0.00564283]" snapshot_42="~[0.0266704,0.00563414]" snapshot_43="~[0.0266702,0.00562553]" snapshot_44="~[0.0266698,0.00561699]" snapshot_45="~[0.0266696,0.00560824]" snapshot_46="~[0.0266693,0.00559965]" snapshot_47="~[0.0266689,0.00559112]" snapshot_48="~[0.0266684,0.00558247]" snapshot_49="~[0.0266679,0.00557379]" snapshot_50="~[0.0266672,0.00556517]" snapshot_51="~[0.0266666,0.00555649]" snapshot_52="~[0.0266662,0.00554791]" snapshot_53="~[0.0266656,0.00553924]" snapshot_54="~[0.0266649,0.00553042]" snapshot_55="~[0.0266644,0.00552183]" snapshot_56="~[0.0266641,0.00551319]" snapshot_57="~[0.0266637,0.00550443]" snapshot_58="~[0.0266633,0.00549567]" snapshot_59="~[0.026663,0.00548677]" snapshot_60="~[0.0266626,0.00547773]" snapshot_61="~[0.0266621,0.00546882]" snapshot_62="~[0.0266615,0.00545986]" snapshot_63="~[0.0266609,0.00545087]" snapshot_64="~[0.0266604,0.00544198]" snapshot_65="~[0.0266599,0.00543305]" snapshot_66="~[0.0266597,0.00542406]" snapshot_67="~[0.0266592,0.00541506]" snapshot_68="~[0.0266588,0.00540594]" snapshot_69="~[0.0266584,0.00539681]" snapshot_70="~[0.0266578,0.00538762]" snapshot_71="~[0.0266572,0.00537829]" snapshot_72="~[0.0266567,0.0053692]" snapshot_73="~[0.0266563,0.00536003]" snapshot_74="~[0.0266556,0.00535068]" snapshot_75="~[0.0266552,0.00534141]" snapshot_76="~[0.0266547,0.00533186]" snapshot_77="~[0.0266542,0.00532239]" snapshot_78="~[0.0266538,0.00531292]" snapshot_79="~[0.0266533,0.00530336]" snapshot_80="~[0.0266528,0.00529358]" snapshot_81="~[0.0266525,0.00528393]" snapshot_82="~[0.0266517,0.00527397]" snapshot_83="~[0.0266514,0.00526431]" snapshot_84="~[0.0266507,0.00525419]" snapshot_85="~[0.0266501,0.00524421]" snapshot_86="~[0.0266496,0.00523413]" snapshot_87="~[0.0266493,0.00522387]" snapshot_88="~[0.0266488,0.0052137]" snapshot_89="~[0.0266484,0.00520334]" snapshot_90="~[0.0266481,0.00519295]" snapshot_91="~[0.0266475,0.00518226]" snapshot_92="~[0.0266468,0.00517162]" snapshot_93="~[0.0266463,0.00516067]" snapshot_94="~[0.0266456,0.00514959]" snapshot_95="~[0.026645,0.00513834]" snapshot_96="~[0.0266446,0.00512726]" snapshot_97="~[0.0266442,0.00511606]" snapshot_98="~[0.0266438,0.00510477]" snapshot_99="~[0.0266435,0.00509322]" snapshot_100="~[0.0266431,0.00508163]" snapshot_101="~[0.0266424,0.00506955]" snapshot_102="~[0.0266419,0.00505755]" snapshot_103="~[0.0266413,0.0050451]" snapshot_104="~[0.0266406,0.00503257]" snapshot_105="~[0.0266402,0.00502035]" snapshot_106="~[0.0266398,0.00500734]" snapshot_107="~[0.0266393,0.00499456]" snapshot_108="~[0.0266388,0.00498163]" snapshot_109="~[0.0266383,0.00496857]" snapshot_110="~[0.0266378,0.00495559]" snapshot_111="~[0.0266373,0.00494242]" snapshot_112="~[0.0266366,0.00492869]" snapshot_113="~[0.0266359,0.00491458]" snapshot_114="~[0.0266352,0.00490047]" snapshot_115="~[0.0266347,0.0048866]" snapshot_116="~[0.0266342,0.0048725]" snapshot_117="~[0.0266338,0.00485836]" snapshot_118="~[0.0266333,0.00484404]" snapshot_119="~[0.0266327,0.00482982]" snapshot_120="~[0.026632,0.00481472]" snapshot_121="~[0.0266312,0.0047999]" snapshot_122="~[0.0266304,0.00478466]" snapshot_123="~[0.0266295,0.00476948]" snapshot_124="~[0.0266291,0.00475451]" snapshot_125="~[0.0266286,0.00473898]" snapshot_126="~[0.0266279,0.00472352]" snapshot_127="~[0.026627,0.00470805]" snapshot_128="~[0.0266264,0.00469234]" snapshot_129="~[0.0266258,0.00467678]" snapshot_130="~[0.0266249,0.00466082]" snapshot_131="~[0.0266244,0.00464504]" snapshot_132="~[0.0266236,0.00462851]" snapshot_133="~[0.0266227,0.00461249]" snapshot_134="~[0.0266219,0.00459607]" snapshot_135="~[0.0266212,0.00457963]" snapshot_136="~[0.0266203,0.004563]" snapshot_137="~[0.0266192,0.00454583]" snapshot_138="~[0.0266183,0.00452889]" snapshot_139="~[0.0266174,0.00451132]" snapshot_140="~[0.0266164,0.00449428]" snapshot_141="~[0.0266153,0.00447682]" snapshot_142="~[0.0266145,0.00445991]" snapshot_143="~[0.0266136,0.00444257]" snapshot_144="~[0.0266127,0.00442523]" snapshot_145="~[0.0266119,0.00440789]" snapshot_146="~[0.0266108,0.00439062]" snapshot_147="~[0.0266099,0.0043734]" snapshot_148="~[0.0266087,0.00435553]" snapshot_149="~[0.0266074,0.00433742]" snapshot_150="~[0.026606,0.00431897]" snapshot_151="~[0.0266048,0.00430147]" snapshot_152="~[0.0266035,0.00428379]" snapshot_153="~[0.0266022,0.00426558]" snapshot_154="~[0.0266008,0.00424768]" snapshot_155="~[0.0265993,0.00422889]" snapshot_156="~[0.0265979,0.0042108]" snapshot_157="~[0.0265967,0.00419306]" snapshot_158="~[0.026595,0.00417441]" snapshot_159="~[0.0265932,0.0041555]" snapshot_160="~[0.0265917,0.0041375]" snapshot_161="~[0.0265901,0.00411876]" snapshot_162="~[0.0265884,0.00410048]" snapshot_163="~[0.026587,0.00408227]" snapshot_164="~[0.0265855,0.00406385]" snapshot_165="~[0.0265838,0.0040452]" snapshot_166="~[0.0265819,0.00402659]" snapshot_167="~[0.0265799,0.00400758]" snapshot_168="~[0.0265781,0.003989]" snapshot_169="~[0.0265762,0.00397019]" snapshot_170="~[0.0265742,0.00395153]" snapshot_171="~[0.0265722,0.00393215]" snapshot_172="~[0.0265696,0.00391303]" snapshot_173="~[0.0265673,0.00389381]" snapshot_174="~[0.0265654,0.00387534]" snapshot_175="~[0.0265634,0.00385656]" snapshot_176="~[0.0265612,0.00383743]" snapshot_177="~[0.0265591,0.00381873]" snapshot_178="~[0.0265566,0.00379921]" snapshot_179="~[0.0265542,0.00377985]" snapshot_180="~[0.0265516,0.00376061]" snapshot_181="~[0.0265492,0.00374198]" snapshot_182="~[0.0265465,0.00372259]" snapshot_183="~[0.0265438,0.00370326]" snapshot_184="~[0.026541,0.0036836]" snapshot_185="~[0.0265382,0.00366429]" snapshot_186="~[0.0265355,0.00364544]" snapshot_187="~[0.0265327,0.00362702]" snapshot_188="~[0.0265297,0.0036082]" snapshot_189="~[0.0265264,0.00358928]" snapshot_190="~[0.0265232,0.00357055]" snapshot_191="~[0.0265199,0.00355209]" snapshot_192="~[0.0265163,0.00353341]" snapshot_193="~[0.0265129,0.00351507]" snapshot_194="~[0.0265092,0.00349668]" snapshot_195="~[0.0265057,0.0034792]" snapshot_196="~[0.0265018,0.00346078]" snapshot_197="~[0.0264978,0.00344285]" />
+    </Result_Element>
+</result>
