+++ tests/2d_examples/test_2d_stretching/regression_test_tool/BeamObserver_Position_Run_4_result.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<result>
+    <Snapshot_Element>
+        <Snapshot number_of_snapshot_for_local_result_="199" />
+    </Snapshot_Element>
+    <Result_Element>
+        <Particle_0 snapshot_0="~[0.0266654,0.0060265]" snapshot_1="~[0.0266664,0.00602312]" snapshot_2="~[0.0266667,0.00601772]" snapshot_3="~[0.0266699,0.0060033]" snapshot_4="~[0.0266712,0.00599011]" snapshot_5="~[0.0266722,0.00597808]" snapshot_6="~[0.0266722,0.00596659]" snapshot_7="~[0.0266725,0.00595593]" snapshot_8="~[0.0266723,0.00594538]" snapshot_9="~[0.0266726,0.00593538]" snapshot_10="~[0.0266727,0.00592551]" snapshot_11="~[0.0266728,0.00591574]" snapshot_12="~[0.0266728,0.00590619]" snapshot_13="~[0.0266729,0.00589688]" snapshot_14="~[0.0266726,0.0058876]" snapshot_15="~[0.0266725,0.00587853]" snapshot_16="~[0.0266725,0.00586949]" snapshot_17="~[0.0266724,0.00586063]" snapshot_18="~[0.0266722,0.00585198]" snapshot_19="~[0.0266719,0.00584302]" snapshot_20="~[0.0266718,0.00583428]" snapshot_21="~[0.0266718,0.00582546]" snapshot_22="~[0.0266713,0.00581671]" snapshot_23="~[0.0266711,0.00580802]" snapshot_24="~[0.0266706,0.00579931]" snapshot_25="~[0.0266703,0.00579058]" snapshot_26="~[0.0266699,0.00578183]" snapshot_27="~[0.0266696,0.0057732]" snapshot_28="~[0.0266693,0.00576445]" snapshot_29="~[0.0266686,0.00575572]" snapshot_30="~[0.0266683,0.00574715]" snapshot_31="~[0.0266675,0.00573856]" snapshot_32="~[0.0266672,0.00572999]" snapshot_33="~[0.0266668,0.00572134]" snapshot_34="~[0.0266662,0.00571275]" snapshot_35="~[0.026666,0.0057042]" snapshot_36="~[0.0266656,0.00569548]" snapshot_37="~[0.026665,0.00568693]" snapshot_38="~[0.0266642,0.00567844]" snapshot_39="~[0.0266637,0.00566986]" snapshot_40="~[0.026663,0.00566125]" snapshot_41="~[0.0266622,0.00565287]" snapshot_42="~[0.0266618,0.00564431]" snapshot_43="~[0.0266613,0.00563573]" snapshot_44="~[0.0266608,0.00562716]" snapshot_45="~[0.0266604,0.00561855]" snapshot_46="~[0.0266599,0.00560994]" snapshot_47="~[0.0266593,0.00560155]" snapshot_48="~[0.0266587,0.00559295]" snapshot_49="~[0.0266581,0.00558448]" snapshot_50="~[0.0266573,0.00557588]" snapshot_51="~[0.0266565,0.00556733]" snapshot_52="~[0.026656,0.0055588]" snapshot_53="~[0.0266554,0.00555024]" snapshot_54="~[0.0266545,0.00554154]" snapshot_55="~[0.026654,0.00553296]" snapshot_56="~[0.0266537,0.00552417]" snapshot_57="~[0.0266531,0.00551548]" snapshot_58="~[0.0266526,0.00550678]" snapshot_59="~[0.0266522,0.00549802]" snapshot_60="~[0.0266516,0.00548927]" snapshot_61="~[0.026651,0.00548036]" snapshot_62="~[0.0266503,0.00547153]" snapshot_63="~[0.0266495,0.00546272]" snapshot_64="~[0.026649,0.00545385]" snapshot_65="~[0.0266485,0.00544495]" snapshot_66="~[0.0266482,0.00543611]" snapshot_67="~[0.0266476,0.00542714]" snapshot_68="~[0.0266471,0.00541802]" snapshot_69="~[0.0266466,0.00540901]" snapshot_70="~[0.026646,0.00540002]" snapshot_71="~[0.0266453,0.0053908]" snapshot_72="~[0.0266448,0.00538164]" snapshot_73="~[0.0266444,0.0053724]" snapshot_74="~[0.0266437,0.00536306]" snapshot_75="~[0.0266434,0.00535378]" snapshot_76="~[0.0266428,0.00534427]" snapshot_77="~[0.0266422,0.00533482]" snapshot_78="~[0.0266419,0.00532531]" snapshot_79="~[0.0266414,0.00531584]" snapshot_80="~[0.026641,0.00530617]" snapshot_81="~[0.0266408,0.00529652]" snapshot_82="~[0.0266401,0.00528651]" snapshot_83="~[0.0266398,0.00527692]" snapshot_84="~[0.0266392,0.00526679]" snapshot_85="~[0.0266388,0.00525672]" snapshot_86="~[0.0266384,0.00524659]" snapshot_87="~[0.0266382,0.00523632]" snapshot_88="~[0.0266379,0.00522607]" snapshot_89="~[0.0266377,0.00521558]" snapshot_90="~[0.0266374,0.00520509]" snapshot_91="~[0.0266371,0.00519431]" snapshot_92="~[0.0266367,0.00518354]" snapshot_93="~[0.0266365,0.00517252]" snapshot_94="~[0.0266361,0.00516119]" snapshot_95="~[0.0266357,0.00514994]" snapshot_96="~[0.0266357,0.00513869]" snapshot_97="~[0.0266356,0.00512731]" snapshot_98="~[0.0266355,0.00511584]" snapshot_99="~[0.0266355,0.00510411]" snapshot_100="~[0.0266355,0.00509235]" snapshot_101="~[0.0266352,0.00508006]" snapshot_102="~[0.0266352,0.00506784]" snapshot_103="~[0.0266351,0.00505535]" snapshot_104="~[0.026635,0.00504275]" snapshot_105="~[0.0266351,0.00503037]" snapshot_106="~[0.0266352,0.00501732]" snapshot_107="~[0.0266354,0.00500436]" snapshot_108="~[0.0266356,0.00499117]" snapshot_109="~[0.0266359,0.00497782]" snapshot_110="~[0.0266361,0.00496443]" snapshot_111="~[0.0266364,0.00495107]" snapshot_112="~[0.0266366,0.00493717]" snapshot_113="~[0.0266367,0.00492285]" snapshot_114="~[0.0266369,0.00490869]" snapshot_115="~[0.0266373,0.00489461]" snapshot_116="~[0.0266377,0.00488039]" snapshot_117="~[0.0266382,0.0048661]" snapshot_118="~[0.0266387,0.0048515]" snapshot_119="~[0.0266392,0.00483711]" snapshot_120="~[0.0266395,0.00482182]" snapshot_121="~[0.0266399,0.00480684]" snapshot_122="~[0.0266403,0.00479123]" snapshot_123="~[0.0266407,0.0047756]" snapshot_124="~[0.0266414,0.00476055]" snapshot_125="~[0.0266421,0.00474494]" snapshot_126="~[0.0266427,0.00472916]" snapshot_127="~[0.0266432,0.00471341]" snapshot_128="~[0.026644,0.00469732]" snapshot_129="~[0.0266448,0.00468141]" snapshot_130="~[0.0266453,0.00466514]" snapshot_131="~[0.0266462,0.00464915]" snapshot_132="~[0.0266469,0.0046324]" snapshot_133="~[0.0266476,0.00461612]" snapshot_134="~[0.0266483,0.00459935]" snapshot_135="~[0.0266492,0.00458268]" snapshot_136="~[0.0266499,0.00456578]" snapshot_137="~[0.0266506,0.00454839]" snapshot_138="~[0.0266514,0.00453103]" snapshot_139="~[0.0266523,0.0045135]" snapshot_140="~[0.0266531,0.00449607]" snapshot_141="~[0.026654,0.00447835]" snapshot_142="~[0.026655,0.00446119]" snapshot_143="~[0.026656,0.00444363]" snapshot_144="~[0.0266571,0.00442593]" snapshot_145="~[0.0266582,0.00440837]" snapshot_146="~[0.0266592,0.00439087]" snapshot_147="~[0.0266604,0.00437321]" snapshot_148="~[0.0266615,0.00435506]" snapshot_149="~[0.0266624,0.00433694]" snapshot_150="~[0.0266633,0.00431811]" snapshot_151="~[0.0266645,0.00430012]" snapshot_152="~[0.0266657,0.00428203]" snapshot_153="~[0.0266668,0.00426339]" snapshot_154="~[0.0266681,0.0042448]" snapshot_155="~[0.0266691,0.00422575]" snapshot_156="~[0.0266704,0.00420732]" snapshot_157="~[0.0266718,0.00418879]" snapshot_158="~[0.0266727,0.00416995]" snapshot_159="~[0.0266738,0.00415071]" snapshot_160="~[0.026675,0.00413224]" snapshot_161="~[0.0266763,0.00411316]" snapshot_162="~[0.0266776,0.00409446]" snapshot_163="~[0.0266792,0.00407575]" snapshot_164="~[0.0266808,0.00405656]" snapshot_165="~[0.0266821,0.0040373]" snapshot_166="~[0.0266833,0.00401805]" snapshot_167="~[0.0266845,0.00399838]" snapshot_168="~[0.026686,0.00397938]" snapshot_169="~[0.0266873,0.00396002]" snapshot_170="~[0.0266886,0.00394072]" snapshot_171="~[0.0266899,0.00392085]" snapshot_172="~[0.0266907,0.00390098]" snapshot_173="~[0.0266916,0.00388132]" snapshot_174="~[0.026693,0.00386248]" snapshot_175="~[0.0266945,0.00384302]" snapshot_176="~[0.0266958,0.00382308]" snapshot_177="~[0.0266971,0.00380377]" snapshot_178="~[0.0266981,0.00378381]" snapshot_179="~[0.0266992,0.00376425]" snapshot_180="~[0.0267003,0.00374455]" snapshot_181="~[0.0267015,0.00372557]" snapshot_182="~[0.0267027,0.00370565]" snapshot_183="~[0.0267037,0.00368584]" snapshot_184="~[0.0267048,0.00366595]" snapshot_185="~[0.0267061,0.00364632]" snapshot_186="~[0.0267074,0.00362695]" snapshot_187="~[0.0267087,0.00360756]" snapshot_188="~[0.0267098,0.00358817]" snapshot_189="~[0.0267106,0.00356832]" snapshot_190="~[0.0267116,0.00354891]" snapshot_191="~[0.0267124,0.00352971]" snapshot_192="~[0.0267133,0.00351]" snapshot_193="~[0.0267141,0.00349065]" snapshot_194="~[0.0267149,0.00347086]" snapshot_195="~[0.0267158,0.0034524]" snapshot_196="~[0.0267163,0.00343296]" snapshot_197="~[0.0267169,0.00341379]" snapshot_198="~[0.0267172,0.00339448]" />
+    </Result_Element>
+</result>
