+++ tests/2d_examples/test_2d_stretching/regression_test_tool/BeamObserver_Position_dtwdistance.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <dtw_distance>
-    <DTWDistance Position_0="7.5183066597939338e-05" />
+    <DTWDistance Position_0="0.0032962085612687326" />
 </dtw_distance>
