+++ tests/2d_examples/test_2d_stretching/regression_test_tool/BeamObserver_Position_runtimes.dat
@@ -1,0 +1,3 @@
+true
+6
+4
