+++ tests/2d_examples/test_2d_viscous_cream_drop/regression_test_tool/CreamObserver_Position_Run_3_result.xml
