+++ tests/3d_examples/test_3d_taylor_bar/regression_test_tool/MyObserver_Position.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <Position>
-    <Snapshot_0 Position_0="~[0.00275652,-3.08681e-07,0.00367062]" />
-    <Snapshot_200 Position_0="~[0.00275652,-3.08681e-07,0.00181116]" />
-    <Snapshot_400 Position_0="~[0.00377447,-6.16299e-06,0.00113409]" />
-    <Snapshot_600 Position_0="~[0.00470439,-3.8558e-06,0.00106379]" />
-    <Snapshot_800 Position_0="~[0.0052266,-5.90968e-06,0.000970699]" />
-    <Snapshot_1000 Position_0="~[0.00555656,-1.3226e-05,0.000984015]" />
-    <Snapshot_1200 Position_0="~[0.00572109,-1.66157e-05,0.000983351]" />
-    <Snapshot_1400 Position_0="~[0.00576604,-2.11768e-05,0.000972748]" />
-    <Snapshot_1600 Position_0="~[0.00579725,-2.41937e-05,0.000986915]" />
-    <Snapshot_1800 Position_0="~[0.00580938,-2.29334e-05,0.000991078]" />
-    <Snapshot_2000 Position_0="~[0.00582087,-2.50791e-05,0.00099723]" />
-    <Snapshot_2200 Position_0="~[0.00582526,-2.69742e-05,0.00101207]" />
+    <Snapshot_0 Position_0="~[-8.10018e-07,1.1244e-06,0.0323979]" />
+    <Snapshot_200 Position_0="~[-8.10018e-07,1.1244e-06,0.0305384]" />
+    <Snapshot_400 Position_0="~[-8.10019e-07,1.1244e-06,0.0286877]" />
+    <Snapshot_600 Position_0="~[-7.7462e-07,1.11762e-06,0.0269074]" />
+    <Snapshot_800 Position_0="~[-7.32718e-07,1.2944e-06,0.0252411]" />
+    <Snapshot_1000 Position_0="~[-7.00057e-07,1.37986e-06,0.0237301]" />
+    <Snapshot_1200 Position_0="~[-5.64912e-07,1.35487e-06,0.0223799]" />
+    <Snapshot_1400 Position_0="~[-3.8431e-07,1.45093e-06,0.0212193]" />
+    <Snapshot_1600 Position_0="~[-3.73306e-07,1.91765e-06,0.0202846]" />
+    <Snapshot_1800 Position_0="~[-1.44953e-07,2.0505e-06,0.0196171]" />
+    <Snapshot_2000 Position_0="~[-1.04484e-07,2.37201e-06,0.0192622]" />
+    <Snapshot_2200 Position_0="~[-9.14834e-08,2.92115e-06,0.0192491]" />
 </Position>
