+++ tests/3d_examples/test_3d_taylor_bar/regression_test_tool/MyObserver_Position_Run_0_result.xml
@@ -4,6 +4,6 @@
         <Snapshot number_of_snapshot_for_local_result_="12" />
     </Snapshot_Element>
     <Result_Element>
-        <Particle_0 snapshot_0="~[0.00271966,-3.92121e-06,0.00369431]" snapshot_1="~[0.00271966,-3.92121e-06,0.00183485]" snapshot_2="~[0.00373463,2.85431e-06,0.0011427]" snapshot_3="~[0.00464875,6.56683e-06,0.00105376]" snapshot_4="~[0.00515902,-4.61543e-06,0.000945498]" snapshot_5="~[0.00551768,-1.28565e-05,0.00101107]" snapshot_6="~[0.00568851,-1.72082e-05,0.00100762]" snapshot_7="~[0.00572551,-1.82997e-05,0.000976733]" snapshot_8="~[0.00575808,-1.59725e-05,0.00100755]" snapshot_9="~[0.00576938,-1.65128e-05,0.00100642]" snapshot_10="~[0.00578132,-1.87098e-05,0.00100775]" snapshot_11="~[0.00578535,-2.13162e-05,0.00102999]" />
+        <Particle_0 snapshot_0="~[-1.99012e-06,-4.38812e-07,0.0323978]" snapshot_1="~[-1.99012e-06,-4.38812e-07,0.0305384]" snapshot_2="~[-1.99012e-06,-4.38811e-07,0.0286872]" snapshot_3="~[-2.0047e-06,-4.4168e-07,0.0269067]" snapshot_4="~[-2.0243e-06,-4.14104e-07,0.0252401]" snapshot_5="~[-2.09274e-06,-3.33955e-07,0.0237295]" snapshot_6="~[-1.69427e-06,-4.39097e-07,0.0223788]" snapshot_7="~[-1.69325e-06,-3.6255e-07,0.0212183]" snapshot_8="~[-1.56376e-06,-5.12507e-07,0.0202828]" snapshot_9="~[-1.58657e-06,-7.99528e-07,0.0196141]" snapshot_10="~[-1.65933e-06,-8.92709e-07,0.019256]" snapshot_11="~[-1.44635e-06,-1.76574e-06,0.0192393]" />
     </Result_Element>
 </result>
