+++ tests/3d_examples/test_3d_taylor_bar/regression_test_tool/MyObserver_Position_Run_2_result.xml
@@ -4,6 +4,6 @@
         <Snapshot number_of_snapshot_for_local_result_="12" />
     </Snapshot_Element>
     <Result_Element>
-        <Particle_0 snapshot_0="~[0.00274099,-2.22788e-06,0.00367842]" snapshot_1="~[0.00274099,-2.22788e-06,0.00181896]" snapshot_2="~[0.00375147,-9.57617e-06,0.0011276]" snapshot_3="~[0.00468131,-1.97498e-05,0.00105821]" snapshot_4="~[0.00520565,-2.70862e-05,0.000956573]" snapshot_5="~[0.00555345,-2.71041e-05,0.000989583]" snapshot_6="~[0.00572343,-3.03443e-05,0.00099181]" snapshot_7="~[0.0057618,-3.02812e-05,0.000973303]" snapshot_8="~[0.00579815,-3.12753e-05,0.000996506]" snapshot_9="~[0.00581127,-3.77705e-05,0.000995841]" snapshot_10="~[0.00582646,-3.14701e-05,0.0010067]" snapshot_11="~[0.00583058,-3.38134e-05,0.00102478]" />
+        <Particle_0 snapshot_0="~[-1.26594e-06,-1.61909e-06,0.0323978]" snapshot_1="~[-1.26594e-06,-1.61909e-06,0.0305384]" snapshot_2="~[-1.26594e-06,-1.61909e-06,0.0286878]" snapshot_3="~[-1.2607e-06,-1.65851e-06,0.0269077]" snapshot_4="~[-1.59869e-06,-1.54319e-06,0.0252409]" snapshot_5="~[-1.63831e-06,-1.51112e-06,0.0237298]" snapshot_6="~[-1.73333e-06,-1.19488e-06,0.0223793]" snapshot_7="~[-1.84961e-06,-9.48568e-07,0.0212183]" snapshot_8="~[-1.9555e-06,-1.00749e-06,0.020283]" snapshot_9="~[-2.02376e-06,-9.09015e-07,0.0196136]" snapshot_10="~[-1.9767e-06,-9.06003e-07,0.0192551]" snapshot_11="~[-2.74618e-06,-1.20418e-06,0.0192375]" />
     </Result_Element>
 </result>
