+++ tests/3d_examples/test_3d_taylor_bar/regression_test_tool/MyObserver_Position_Run_4_result.xml
@@ -4,6 +4,6 @@
         <Snapshot number_of_snapshot_for_local_result_="12" />
     </Snapshot_Element>
     <Result_Element>
-        <Particle_0 snapshot_0="~[0.00273892,-8.57571e-06,0.00368179]" snapshot_1="~[0.00273892,-8.57571e-06,0.00182233]" snapshot_2="~[0.00375226,-8.84493e-06,0.00113565]" snapshot_3="~[0.00467967,-1.40037e-06,0.00106424]" snapshot_4="~[0.00519828,1.18415e-06,0.000961099]" snapshot_5="~[0.00554447,1.60045e-06,0.000995708]" snapshot_6="~[0.00571541,3.11627e-06,0.00099671]" snapshot_7="~[0.00576645,3.25398e-06,0.000977871]" snapshot_8="~[0.00580604,1.1222e-05,0.000994893]" snapshot_9="~[0.00582163,1.42843e-05,0.000996146]" snapshot_10="~[0.00584076,1.92627e-05,0.0010075]" snapshot_11="~[0.00584533,2.1515e-05,0.0010249]" />
+        <Particle_0 snapshot_0="~[5.66679e-07,-8.51282e-07,0.0323979]" snapshot_1="~[5.66679e-07,-8.51282e-07,0.0305384]" snapshot_2="~[5.6668e-07,-8.51283e-07,0.0286881]" snapshot_3="~[5.80917e-07,-9.17117e-07,0.0269078]" snapshot_4="~[7.43766e-07,-1.19525e-06,0.0252409]" snapshot_5="~[6.82823e-07,-1.59852e-06,0.0237297]" snapshot_6="~[6.70146e-07,-2.19432e-06,0.0223789]" snapshot_7="~[9.7415e-07,-2.91849e-06,0.0212182]" snapshot_8="~[1.23129e-06,-4.17831e-06,0.0202833]" snapshot_9="~[1.47913e-06,-5.71863e-06,0.0196164]" snapshot_10="~[2.0847e-06,-7.95038e-06,0.0192628]" snapshot_11="~[2.40536e-06,-1.0457e-05,0.0192546]" />
     </Result_Element>
 </result>
