+++ tests/3d_examples/test_3d_taylor_bar/taylor_bar.cpp
@@ -172,7 +172,7 @@ int main(int ac, char *av[])
 
     if (sph_system.GenerateRegressionData())
     {
-        write_displacement.generateDataBase(0.08);
+        write_displacement.generateDataBase(0.1);
     }
     else
     {
