+++ tests/3d_examples/test_3d_taylor_bar/taylor_bar.h
@@ -77,6 +77,6 @@ class ColumnObserverParticleGenerator : public ObserverParticleGenerator
   public:
     explicit ColumnObserverParticleGenerator(SPHBody &sph_body) : ObserverParticleGenerator(sph_body)
     {
-        positions_.push_back(Vecd(PL, 0.0, 0.1 * PW));
+        positions_.push_back(Vecd(0.0, 0.0, PW));
     }
 };
